// frontend/src/app/shared/generated-api/model/notebookReorder.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NotebookOrderEntry } from './notebookOrderEntry';


export interface NotebookReorder { 
    notebooks: Array<NotebookOrderEntry>;
}