// frontend/src/app/shared/generated-api/model/notebookUpdate.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContentColor } from './contentColor';


export interface NotebookUpdate { 
    name?: string;
    color?: ContentColor;
    version: number;
}