// frontend/src/app/shared/generated-api/model/noteCreate.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NoteType } from './noteType';
import { ContentColor } from './contentColor';
import { ChecklistItemInput } from './checklistItemInput';


export interface NoteCreate { 
    id?: string;
    notebookId?: string;
    noteType: NoteType;
    title?: string;
    content?: string;
    checklistItems?: Array<ChecklistItemInput>;
    pinned?: boolean;
    archived?: boolean;
    color?: ContentColor;
    labelIds?: Array<string>;
    imageIds?: Array<string>;
}